packages/training/pyproject.toml
[project]
name = "pnseg-training"
version = "0.1.0"
dependencies = [
    "pnseg-core",
    "segmentation-models-pytorch",
    "mlflow",
    "albumentations",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["training*"]